safari-600.7-branch/LayoutTests/platform/ios-sim/fast/dom/navigation-type-back-forward-expected.txt

This tests that performance.navigation.type is 2

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL navigation.type should be 2 (of type number). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
